Components/profiling: add LLVM/Clang -coverage support in gcov.c
- Add LLVM/Clang coverage runtime support with data structures and
  callback functions based on Linux kernel's gcov/clang.c implementation
- Refactor GCC-specific code with `#ifndef __clang__` guards for better
  separation of GCC and LLVM implementations
- Add comprehensive Doxygen-style documentation for all data
  structures and functions
- Fix printf format specifiers in gcov.c and gprof.c for better
  portability (use %lu with explicit casts for pointer values)
- Update file header comment to reflect LLVM/Clang support
- Add changelog entries documenting the new features

Note: LLVM gcov format is incompatible with GCC gcov format due to
different data structures and runtime mechanisms. Projects should
use consistent toolchain (either GCC or LLVM) for coverage analysis.

see https://clang.llvm.org/docs/SourceBasedCodeCoverage.html

This enables code coverage analysis when compiling with LLVM/Clang
using the -coverage option, expanding beyond the previous GCC-only
support.

+++ Components/profiling/gprof.c
@@ -499,7 +499,7 @@ long gprof_collect(unsigned long interface)
     }
     if (interface == 0) {
         gprof_data.size = bufptr - gprof_data.buf;
-        printf("Collected gprof data @0x%lx, size %lu bytes\n", gprof_data.buf, gprof_data.size);
+        printf("Collected gprof data @0x%lx, size %u bytes\n", (unsigned long)(gprof_data.buf), gprof_data.size);
     } else if (interface == 1) {
         fclose(fp);
         printf("Write %s done!\n", gmon_out);
